Upgrading to a newer version of the SDK
