+++ src/apdu/dispatcher.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "parser.h"
+
 #include "../types.h"
 
 /**
