+++ src/handler/get_public_key.c
@@ -22,14 +22,14 @@
 
 #include "os.h"
 #include "cx.h"
+#include "io.h"
+#include "buffer.h"
 
 #include "get_public_key.h"
 #include "../globals.h"
 #include "../types.h"
-#include "../io.h"
 #include "../sw.h"
 #include "../crypto.h"
-#include "../common/buffer.h"
 #include "../ui/display.h"
 #include "../helper/send_response.h"
