+++ src/handler/get_public_key.h
@@ -4,8 +4,9 @@
 #include <stdbool.h>  // bool
 #include <stdint.h>   // uint*_t
 
+#include "buffer.h"
+
 #include "../types.h"
-#include "../common/buffer.h"
 
 /**
  * Handler for GET_PUBLIC_KEY command. If successfully parse BIP32 path,
