+++ src/handler/get_version.c
@@ -19,13 +19,14 @@
 #include <limits.h>  // UINT8_MAX
 #include <assert.h>  // _Static_assert
 
+#include "io.h"
+#include "buffer.h"
+
 #include "get_version.h"
 #include "../globals.h"
 #include "../constants.h"
-#include "../io.h"
 #include "../sw.h"
 #include "../types.h"
-#include "common/buffer.h"
 
 int handler_get_version() {
     _Static_assert(APPVERSION_LEN == 3, "Length of (MAJOR || MINOR || PATCH) must be 3!");
@@ -36,11 +37,10 @@ int handler_get_version() {
     _Static_assert(PATCH_VERSION >= 0 && PATCH_VERSION <= UINT8_MAX,
                    "PATCH version must be between 0 and 255!");
 
-    return io_send_response(
-        &(const buffer_t){.ptr = (uint8_t[APPVERSION_LEN]){(uint8_t) MAJOR_VERSION,
-                                                           (uint8_t) MINOR_VERSION,
-                                                           (uint8_t) PATCH_VERSION},
-                          .size = APPVERSION_LEN,
-                          .offset = 0},
+    return io_send_response_pointer(
+        (const uint8_t *) &(uint8_t[APPVERSION_LEN]){(uint8_t) MAJOR_VERSION,
+                                                     (uint8_t) MINOR_VERSION,
+                                                     (uint8_t) PATCH_VERSION},
+        APPVERSION_LEN,
         SW_OK);
 }
