+++ src/handler/sign_tx.c
@@ -22,13 +22,13 @@
 
 #include "os.h"
 #include "cx.h"
+#include "buffer.h"
 
 #include "sign_tx.h"
 #include "../sw.h"
 #include "../globals.h"
 #include "../crypto.h"
 #include "../ui/display.h"
-#include "../common/buffer.h"
 #include "../transaction/types.h"
 #include "../transaction/deserialize.h"
