+++ src/helper/send_reponse.c
@@ -19,11 +19,12 @@
 #include <stdint.h>  // uint*_t
 #include <string.h>  // memmove
 
+#include "buffer.h"
+
 #include "send_response.h"
 #include "../constants.h"
 #include "../globals.h"
 #include "../sw.h"
-#include "common/buffer.h"
 
 int helper_send_response_pubkey() {
     uint8_t resp[1 + 1 + PUBKEY_LEN + 1 + CHAINCODE_LEN] = {0};
@@ -37,7 +38,7 @@ int helper_send_response_pubkey() {
     memmove(resp + offset, G_context.pk_info.chain_code, CHAINCODE_LEN);
     offset += CHAINCODE_LEN;
 
-    return io_send_response(&(const buffer_t){.ptr = resp, .size = offset, .offset = 0}, SW_OK);
+    return io_send_response_pointer(resp, offset, SW_OK);
 }
 
 int helper_send_response_sig() {
@@ -48,5 +49,5 @@ int helper_send_response_sig() {
     memmove(resp + offset, G_context.tx_info.signature, G_context.tx_info.signature_len);
     offset += G_context.tx_info.signature_len;
 
-    return io_send_response(&(const buffer_t){.ptr = resp, .size = offset, .offset = 0}, SW_OK);
+    return io_send_response_pointer(resp, offset, SW_OK);
 }
