--- src/offsets.h
@@ -1,26 +1,0 @@
-#pragma once
-
-/**
- * Offset of instruction class.
- */
-#define OFFSET_CLA 0
-/**
- * Offset of instruction code.
- */
-#define OFFSET_INS 1
-/**
- * Offset of instruction parameter 1.
- */
-#define OFFSET_P1 2
-/**
- * Offset of instruction parameter 2.
- */
-#define OFFSET_P2 3
-/**
- * Offset of command data length.
- */
-#define OFFSET_LC 4
-/**
- * Offset of command data.
- */
-#define OFFSET_CDATA 5
