+++ src/transaction/deserialize.h
@@ -1,7 +1,8 @@
 #pragma once
 
+#include "buffer.h"
+
 #include "types.h"
-#include "../common/buffer.h"
 
 /**
  * Deserialize raw transaction in structure.
