+++ src/types.h
@@ -3,18 +3,10 @@
 #include <stddef.h>  // size_t
 #include <stdint.h>  // uint*_t
 
+#include "bip32.h"
+
 #include "constants.h"
 #include "transaction/types.h"
-#include "common/bip32.h"
-
-/**
- * Enumeration for the status of IO.
- */
-typedef enum {
-    READY,     /// ready for new event
-    RECEIVED,  /// data received
-    WAITING    /// waiting
-} io_state_e;
 
 /**
  * Enumeration with expected INS of APDU commands.
@@ -25,18 +17,6 @@ typedef enum {
     GET_PUBLIC_KEY = 0x05,  /// public key of corresponding BIP32 path
     SIGN_TX = 0x06          /// sign transaction with BIP32 path
 } command_e;
-
-/**
- * Structure with fields of APDU command.
- */
-typedef struct {
-    uint8_t cla;    /// Instruction class
-    command_e ins;  /// Instruction code
-    uint8_t p1;     /// Instruction parameter 1
-    uint8_t p2;     /// Instruction parameter 2
-    uint8_t lc;     /// Length of command data
-    uint8_t *data;  /// Command data
-} command_t;
 
 /**
  * Enumeration with parsing state.
